server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pymongo==4.7.0
boto3==1.34.162
httpx==0.27.0
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
pytesseract==0.3.10

# local
mongomock==4.1.2
python-multipart==0.0.9